posture_backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
opencv-python==4.8.1.78
mediapipe==0.10.7
numpy==1.24.3
pillow==10.0.1
pandas==2.0.3
scikit-learn==1.3.0
tensorflow==2.13.0
requests==2.31.0